requirements.txt
numpy
iio==0.0.3
openai-whisper
